Update plugin file with binarylane changes
This commit is the result of squashing a lot of work in to one package.

The BL class can now be imported, instantiated, and all its methods return
valid data (when provided valid inputs).

The Authenticatoin class talks to ACME servers and calls BL at the right
moment to add and remove records.